requirements.txt
asgiref==3.3.0
Django==3.1.3
django-braces==1.14.0
django-cors-headers==3.6.0
django-crispy-forms==1.9.2
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1

